Modified API and library backend to accomdate single-sequence devices
Not all devices support multiple test sequences. Almost all support multiple test steps however.
Also wanted to make the backend less redundant and more flexible to different generations of
command sets. Now the test editor itself will serve as object code for the test and then a
separate compiler object will function as an iterator that produces byte buffers which can be
sent to the device. This is opposed to the current approach of turning into a rigid representation
of the device commands/code as a Rust enum type.