Implemented multiple display/compilation modes for the command set Made the command set branch into several sub-enums for clarity of which parameters apply to which tests Decided to waylay zero allocations work -- too time consuming right now. Just need something working
Idea in this regard though: the builders can use the new sub-enums now and use a fixed size array
where the parameters are at constant offsets into the array. This would make compilation very
simple to implement and probably allow skipping it as an explicit step. Make it an iterator
instead